add midleware

+++ internal/service/documentService.go
@@ -1,13 +1,16 @@
 package service
 
 import (
+	"errors"
 	"fmt"
-	"github.com/Mallbrusss/BackEntryMiddle/models"
-	"github.com/Mallbrusss/BackEntryMiddle/internal/repository"
 	"mime"
 	"os"
 	"path/filepath"
+	"sort"
 	"sync"
+
+	"github.com/Mallbrusss/BackEntryMiddle/internal/repository"
+	"github.com/Mallbrusss/BackEntryMiddle/models"
 
 	"github.com/google/uuid"
 )
@@ -71,10 +74,10 @@ func (ds *DocumentService) UploadDocument(document *models.Document, fileData []
 	}
 
 	if fError != nil {
-			removeErr := os.Remove(filePath)
-			if removeErr != nil && !os.IsNotExist(removeErr) {
-				return nil, fmt.Errorf("error: %v, error deleting file: %v", fError, removeErr)
-			}
+		removeErr := os.Remove(filePath)
+		if removeErr != nil && !os.IsNotExist(removeErr) {
+			return nil, fmt.Errorf("error: %v, error deleting file: %v", fError, removeErr)
+		}
 		return nil, fError
 	}
 
@@ -89,9 +92,54 @@ func (ds *DocumentService) DeleteDocument(documentID string) error {
 		return err
 	}
 
-	if err := os.Remove(document.FilePath); err != nil && !os.IsExist(err) {
-		return err
+	if document.File && document.FilePath != "" {
+		if err := os.Remove(document.FilePath); err != nil && !os.IsExist(err) {
+			return err
+		}
+
 	}
 
 	return ds.docRepo.DeleteDocument(document)
+}
+
+func (ds *DocumentService) GetDocumentByID(documentID, token string) (*models.Document, error) {
+	document, err := ds.docRepo.GetDocumentByID(documentID, token)
+	if err != nil {
+		return nil, err
+	}
+
+	return document, nil
+
+}
+
+func (ds *DocumentService) GetDocuments(login string, filter map[string]any, limit int) ([]models.Document, error) {
+	documents, err := ds.docRepo.GetDocuments(login, filter, limit)
+	if err != nil {
+		return nil, err
+	}
+
+	sort.Slice(documents, func(i, j int) bool {
+		if documents[i].Name == documents[j].Name {
+			return documents[i].CreatedAt.Before(documents[j].CreatedAt)
+		}
+		return documents[i].Name < documents[j].Name
+	})
+
+	return documents, nil
+}
+
+func (ds *DocumentService) GetUserByToken(token string) (*models.User, error) {
+	session, err := ds.docRepo.FindByToken(token)
+	// if err != nil || session.IsExpired() {
+	//     return nil, errors.New("invalid or expired token")
+	// }
+	if err != nil {
+		return nil, errors.New("invalid or expired token")
+	}
+
+	user, err := ds.docRepo.FindByLogin(session.Login)
+	if err != nil {
+        return nil, err
+    }
+	return user, nil
 }
